NOTICK Current sandbox thread local
Create `CurrentSandboxGroupContextService` to set, get and remove the
current `SandboxGroupContext` on a thread.

This service can then be OSGI injected into other services to retrieve
the current sandbox. The thead local concept is private to the service.
