+++ components/flow/flow-service/src/main/kotlin/net/corda/flow/fiber/FlowFiberExecutionContext.kt
@@ -4,6 +4,7 @@ import net.corda.flow.pipeline.sandbox.FlowSandboxGroupContext
 import net.corda.flow.state.FlowCheckpoint
 import net.corda.flow.state.FlowStack
 import net.corda.membership.read.MembershipGroupReader
+import net.corda.sandboxgroupcontext.SandboxGroupContextService
 import net.corda.serialization.checkpoint.NonSerializable
 import net.corda.v5.base.types.MemberX500Name
 import net.corda.virtualnode.HoldingIdentity
@@ -13,7 +14,8 @@ class FlowFiberExecutionContext(
     val flowCheckpoint: FlowCheckpoint,
     val sandboxGroupContext: FlowSandboxGroupContext,
     val holdingIdentity: HoldingIdentity,
-    val membershipGroupReader: MembershipGroupReader
+    val membershipGroupReader: MembershipGroupReader,
+    val sandboxGroupContextService: SandboxGroupContextService
 ) : NonSerializable {
     val memberX500Name: MemberX500Name = holdingIdentity.x500Name
     val flowStackService: FlowStack = flowCheckpoint.flowStack
