+++ components/persistence/persistence-service-common/src/main/kotlin/net/corda/persistence/common/EntitySandboxServiceFactory.kt
@@ -3,13 +3,13 @@ package net.corda.persistence.common
 import net.corda.cpiinfo.read.CpiInfoReadService
 import net.corda.db.connection.manager.DbConnectionManager
 import net.corda.persistence.common.internal.EntitySandboxServiceImpl
-import net.corda.sandboxgroupcontext.service.SandboxGroupContextComponent
+import net.corda.sandboxgroupcontext.service.SandboxGroupComponent
 import net.corda.virtualnode.read.VirtualNodeInfoReadService
 import org.osgi.service.component.ComponentContext
 
 class EntitySandboxServiceFactory {
     fun create(
-        sandboxService: SandboxGroupContextComponent,
+        sandboxService: SandboxGroupComponent,
         cpiInfoService: CpiInfoReadService,
         virtualNodeInfoService: VirtualNodeInfoReadService,
         dbConnectionManager: DbConnectionManager,
