+++ components/virtual-node/ledger-persistence-service-impl/src/integrationTest/kotlin/net/corda/ledger/persistence/impl/tests/ConsensualLedgerMessageProcessorTests.kt
@@ -213,7 +213,7 @@ class ConsensualLedgerMessageProcessorTests {
         // set up sandbox
         val entitySandboxService =
             EntitySandboxServiceFactory().create(
-                virtualNode.sandboxGroupContextComponent,
+                virtualNode.sandboxGroupComponent,
                 cpiInfoReadService,
                 virtualNodeInfoReadService,
                 dbConnectionManager,
