+++ components/virtual-node/sandbox-group-context-service/src/main/kotlin/net/corda/sandboxgroupcontext/service/SandboxGroupComponent.kt
@@ -1,0 +1,7 @@
+package net.corda.sandboxgroupcontext.service
+
+import net.corda.lifecycle.Lifecycle
+import net.corda.sandboxgroupcontext.SandboxGroupService
+
+interface SandboxGroupComponent : SandboxGroupService, CacheConfiguration, Lifecycle
+
