--- components/virtual-node/sandbox-group-context-service/src/main/kotlin/net/corda/sandboxgroupcontext/service/SandboxGroupContextComponent.kt
@@ -1,7 +1,0 @@
-package net.corda.sandboxgroupcontext.service
-
-import net.corda.lifecycle.Lifecycle
-import net.corda.sandboxgroupcontext.SandboxGroupContextService
-
-interface SandboxGroupContextComponent : SandboxGroupContextService, CacheConfiguration, Lifecycle
-
